service_ia/requirements.txt
fastapi
uvicorn
mysql-connector-python
httpx
python-multipart
python-dotenv